+++ bims/models/__init__.py
@@ -10,3 +10,4 @@ from bims.models.cluster import *  # noqa
 from bims.models.boundary import *  # noqa
 from bims.models.boundary_type import *  # noqa
 from bims.models.carousel_header import CarouselHeader
+from bims.models.links import *
