+++ core/settings/project.py
@@ -14,7 +14,6 @@ from .contrib import *  # noqa
 INSTALLED_APPS += (
     'bims',
     'example',
-    'links',
 )
 
 # Set languages which want to be translated
