--- links/apps.py
@@ -1,10 +1,0 @@
-# coding: utf-8
-__author__ = 'Alison Mukoma <alison@kartoza.com>'
-__license__ = 'GPL'
-__copyright__ = 'kartoza.com'
-
-from django.apps import AppConfig
-
-
-class LinksConfig(AppConfig):
-    name = 'links'
