--- links/views/__init__.py
@@ -1,4 +1,0 @@
-# coding: utf-8
-__author__ = 'Alison Mukoma <alison@kartoza.com>'
-__license__ = 'GPL'
-__copyright__ = 'kartoza.com'
